Fix BUG-931
Only remove existing relation to the komoto/komo that is being
updated (ie. problem was that previously all parent komo child relations were removed)

+++ tarjonta-service/src/main/java/fi/vm/sade/tarjonta/service/impl/conversion/rest/KoulutusDTOConverterToEntity.java
@@ -762,7 +762,10 @@ public class KoulutusDTOConverterToEntity {
         for (String parentKomoOid : parents) {
             Koulutusmoduuli parent = koulutusmoduuliDAO.findByOid(parentKomoOid);
             for (KoulutusSisaltyvyys sisaltyvyys : parent.getSisaltyvyysList()) {
-                koulutusSisaltyvyysDAO.remove(sisaltyvyys);
+                if (sisaltyvyys.getAlamoduuliList().contains(komoto.getKoulutusmoduuli())
+                        && sisaltyvyys.getAlamoduuliList().size() == 1) {
+                    koulutusSisaltyvyysDAO.remove(sisaltyvyys);
+                }
             }
         }
